+++ ui/src/app/core/accessories/types/garagedooropener/garagedooropener.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, Input } from '@angular/core';
 import { ServiceTypeX } from '@/app/core/accessories/accessories.interfaces';
 
 @Component({
@@ -6,12 +6,10 @@ import { ServiceTypeX } from '@/app/core/accessories/accessories.interfaces';
   templateUrl: './garagedooropener.component.html',
   styleUrls: ['./garagedooropener.component.scss'],
 })
-export class GaragedooropenerComponent implements OnInit {
+export class GaragedooropenerComponent {
   @Input() public service: ServiceTypeX;
 
   constructor() {}
-
-  ngOnInit() {}
 
   onClick() {
     this.service.getCharacteristic('TargetDoorState').setValue(this.service.values.TargetDoorState ? 0 : 1);