+++ ui/src/app/modules/status/credits/credits.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
@@ -6,10 +6,8 @@ import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
   templateUrl: './credits.component.html',
   styleUrls: ['./credits.component.scss'],
 })
-export class CreditsComponent implements OnInit {
+export class CreditsComponent {
   constructor(
     public activeModal: NgbActiveModal,
   ) { }
-
-  ngOnInit() {}
 }